refactor: 컬러 경로 수정
