+++ src/main/java/com/pm/projectmanager/domain/category/Category.java
@@ -1,6 +1,6 @@
 package com.pm.projectmanager.domain.category;
 
-import com.pm.projectmanager.domain.project.Color;
+import com.pm.projectmanager.common.Color;
 import com.pm.projectmanager.domain.project.Project;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
