+++ src/main/java/com/pm/projectmanager/domain/category/dto/CreateCategoryRequestDto.java
@@ -1,6 +1,7 @@
 package com.pm.projectmanager.domain.category.dto;
 
-import com.pm.projectmanager.domain.project.Color;
+import com.pm.projectmanager.common.Color;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
