+++ src/main/java/com/pm/projectmanager/domain/category/dto/UpdateCategoryRequestDto.java
@@ -1,6 +1,7 @@
 package com.pm.projectmanager.domain.category.dto;
 
-import com.pm.projectmanager.domain.project.Color;
+import com.pm.projectmanager.common.Color;
+
 import lombok.Getter;
 
 @Getter
